+++ CHANGELOG.md
@@ -32,6 +32,7 @@ v0.18 (trunk)
  * Add support for XEP-0319: Last User Interaction in Presence and refactor the "last seen" feature
  * Protect Ajax calls when session is dropped, return 403 and redirect properly
  * Add an indexed parent column on Info to ensure the component origins
+ * Request all the latest messages in one query instead of a loop in Chats
 
 v0.17.1
 ---------------------------